requirements.txt

torch
transformers
sklearn
flask
flask_cors
python-docx
boto3
